msm8998-common: add a 'start_log_capture' service
This is a bash script that just stores logcat and dmesg in the cache partition.

This is meant to be used ONLY during development to store logs in certain
conditions, for example when the device reboots itself, or when the device
shutdown without allowing the user to get any log.

This requires SELinux in permissive (it should be possible to create rules
to run this in enforcing, but as it is meant to be used only in development...).

The script:
- starts automatically at boot
- waits for the cache partition to be mounted
- starts to get logcat and dmesg
- the file names with the logs contain the current date (this may always be
1970 anyway...)
- if the file already exists, a file_# is created
- if a file gets bigger than 20MB, it is deleted (this may happen with the
logcat, and the cache partition is limited in size).

Even with the script running, the device can enter deep sleep and vmin.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>